chore(core): add useRemoteTransactions hook
