+++ packages/sanity/src/core/store/events/useEventsStore.ts
@@ -2,15 +2,9 @@
 /* eslint-disable max-nested-callbacks */
 import {useCallback, useMemo} from 'react'
 import {useObservable} from 'react-rx'
-import {BehaviorSubject, combineLatest, from, type Observable, of, Subject} from 'rxjs'
+import {combineLatest, from, type Observable, of, Subject} from 'rxjs'
 import {catchError, map, startWith, switchMap, tap} from 'rxjs/operators'
-import {
-  type MendozaPatch,
-  type SanityClient,
-  type SanityDocument,
-  type TransactionLogEventWithEffects,
-  type WithVersion,
-} from 'sanity'
+import {type SanityClient, type SanityDocument} from 'sanity'
 
 import {useClient} from '../../hooks'
 import {useReleasesStore} from '../../releases/store/useReleasesStore'
@@ -18,10 +12,9 @@ import {getReleaseIdFromName} from '../../releases/util/getReleaseIdFromName'
 import {DEFAULT_STUDIO_CLIENT_OPTIONS} from '../../studioClient'
 import {getVersionFromId, isVersionId} from '../../util/draftUtils'
 import {type DocumentVariantType, getDocumentVariantType} from '../../util/getDocumentVariantType'
-import {type DocumentRemoteMutationEvent} from '../_legacy/document/buffered-doc/types'
 import {getDocumentChanges} from './getDocumentChanges'
 import {getDocumentTransactions} from './getDocumentTransactions'
-import {getEditEvents, getEffectState} from './getEditEvents'
+import {getEditEvents} from './getEditEvents'
 import {
   type DocumentGroupEvent,
   type EditDocumentVersionEvent,
@@ -38,7 +31,7 @@ import {
   isUnscheduleDocumentVersionEvent,
   isUpdateLiveDocumentEvent,
 } from './types'
-import {useRemoteMutations} from './useRemoteMutations'
+import {useRemoteTransactions} from './useRemoteTransactions'
 
 const INITIAL_VALUE = {
   events: [],
@@ -122,24 +115,6 @@ const addEventId = (
   return {...event, id} as DocumentGroupEvent
 }
 
-function remoteMutationToTransaction(
-  event: DocumentRemoteMutationEvent,
-): TransactionLogEventWithEffects {
-  return {
-    author: event.author,
-    documentIDs: [],
-    id: event.transactionId,
-    timestamp: event.timestamp.toISOString(),
-    effects: {
-      [event.head._id]: {
-        // TODO: Find a way to validate that is a MendozaPatch
-        apply: event.effects.apply as MendozaPatch,
-        revert: event.effects.revert as MendozaPatch,
-      },
-    },
-  }
-}
-
 export function useEventsStore({
   documentId,
   documentType,
@@ -153,14 +128,19 @@ export function useEventsStore({
 }): EventsStore {
   const client = useClient(DEFAULT_STUDIO_CLIENT_OPTIONS)
   const documentVariantType = getDocumentVariantType(documentId)
-  const remoteTransactions$ = useMemo(
-    () => new BehaviorSubject<TransactionLogEventWithEffects[]>([]),
-    [],
-  )
+
+  const refreshEventsTrigger$ = useMemo(() => new Subject<void>(), [])
+  const refetchEvents = useCallback(() => refreshEventsTrigger$.next(), [refreshEventsTrigger$])
+  const remoteTransactions$ = useRemoteTransactions({
+    client,
+    documentId,
+    documentType,
+    documentVariantType,
+    onRefetch: refetchEvents,
+  })
 
   const {state$} = useReleasesStore()
 
-  const refreshEventsTrigger$ = useMemo(() => new Subject<void>(), [])
   const eventsObservable$: Observable<{
     events: DocumentGroupEvent[]
     nextCursor: string
@@ -252,56 +232,6 @@ export function useEventsStore({
       ),
     )
   }, [client, documentId, refreshEventsTrigger$, documentVariantType])
-
-  const onUpdate = useCallback(
-    (remoteMutation: WithVersion<DocumentRemoteMutationEvent>) => {
-      // If the remote mutation happened to a published document we need to re-fetch the events.
-      // If it happens to a version, we need to add the mutation to the list of events.
-      // If it happens to a draft: we need to decide if it looks like an event
-      //       Looks like an event: we need to refetch the events list (e.g. publish, discard)
-      //       Doesn't look like an event: we need to add the mutation to the list of events.
-      const variant = remoteMutation.version
-      if (variant !== documentVariantType) {
-        // The mutation is not for the current document variant, we don't need to do anything.
-        return
-      }
-      if (variant === 'version') {
-        remoteTransactions$.next([
-          ...remoteTransactions$.value,
-          remoteMutationToTransaction(remoteMutation),
-        ])
-        return
-      }
-      if (variant === 'draft') {
-        const effectState = getEffectState({
-          apply: remoteMutation.effects.apply as MendozaPatch,
-          revert: remoteMutation.effects.revert as MendozaPatch,
-        })
-        if (effectState === 'upsert' || effectState === 'unedited') {
-          remoteTransactions$.next([
-            ...remoteTransactions$.value,
-            remoteMutationToTransaction(remoteMutation),
-          ])
-        } else {
-          refreshEventsTrigger$.next()
-        }
-        return
-      }
-      if (variant === 'published') {
-        refreshEventsTrigger$.next()
-        return
-      }
-      console.error('Unknown variant', variant)
-    },
-    [remoteTransactions$, refreshEventsTrigger$, documentVariantType],
-  )
-
-  useRemoteMutations({
-    client,
-    documentId,
-    documentType,
-    onUpdate: onUpdate,
-  })
 
   const observable$ = useMemo(() => {
     return combineLatest([state$, eventsObservable$, remoteTransactions$]).pipe(
